refactor: Popup size structure
