+++ src/model/PopupSize.ts
@@ -1,0 +1,9 @@
+export type PopupSize = {
+  height: number;
+  width: number;
+};
+
+export const defaultPopupSize = {
+  height: 600,
+  width: 400,
+};
