--- src/model/settings/PopupSize.ts
@@ -1,18 +1,0 @@
-export class PopupSize {
-  constructor(
-    public height: number,
-    public width: number,
-  ) {}
-
-  static default(): PopupSize {
-    return new PopupSize(600, 400);
-  }
-
-  updateHeight(value: number): PopupSize {
-    return new PopupSize(value, this.width);
-  }
-
-  updateWidth(value: number): PopupSize {
-    return new PopupSize(this.height, value);
-  }
-}
